+++ ndonnx/__init__.py
@@ -161,6 +161,7 @@ from ._funcs import (
     take,
     UnsupportedOperationError,
 )
+from ._info import __array_namespace_info__
 from ._constants import (
     e,
     inf,
@@ -176,6 +177,7 @@ except importlib.metadata.PackageNotFoundError as err:  # pragma: no cover
 
 
 __all__ = [
+    "__array_namespace_info__",
     "Array",
     "array",
     "from_spox_var",